extract chat room's logics to its own resource controller
